Refactor SettingsContext to remove updateSettings function and enhance saveSettings for better handling of fixed options; update Settings component to utilize new saveSettings implementation.
